+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/expressions/TreeNode.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.table.planner.plan
+package org.apache.flink.table.planner.expressions
 
 import org.apache.flink.table.planner.typeutils.TypeInfoCheckUtils
